Add backend services with Google Cloud Functions
Implemented all the endpoint logic to send a verification code of 6 digits, create a new account using the user's email and signing in the user by creating a custom token so the app can sign in in the Flutter app. Also created all the unit tests and coverage scripts for the functions.